requirements.txt
Django==4.2.26
djangorestframework==3.15.2
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
channels==4.2.0
celery==5.4.0
redis==5.2.1
